Move code from `tests.rs` to `scan.rs`
